fix: shulker boxes treated as beacon shulker (#3)
The beacon shulker check didn't account for the needed Block NBT tag
that the item in slot 13 had to have to be a valid beacon shulker. Thus,
if a shulker box with an item in slot 13 (the middle one) got placed, it
was treated as if it was a beacon shulker. A beacon would have been
placed and the datapack tried to create the beacon layers. Since the
data for those layers wasn't present, the datapack would simply drop an
empty shulker box, thus deleting all items inside it.

Fixes #2